requirements.txt
matplotlib
numpy
pandas
scikit-learn
scipy
seaborn
jupyter
xgboost
nltk
wordcloud
sklearn
tensorflow-gpu